added some comments etc

+++ src/main/java/se/kth/iv1201/recruitment/domain/Applicant.java
@@ -1,34 +1,48 @@
 package se.kth.iv1201.recruitment.domain;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Pattern;
+import javax.validation.constraints.Size;
 
 /**
  * https://spring.io/guides/gs/accessing-data-jpa/
  */
 @Entity
-// @Table(name = "APPLICANT")
+//@Table(name = "APPLICANT")
 public class Applicant implements ApplicantDTO {
     @Id
-    // @Column(name = "APPLICANT_USERNAME")
+   // @GeneratedValue(strategy = GenerationType.IDENTITY)
+   // @Column(name = "APPLICANT_USERNAME")
     private String username;
 
-    //@Column(name = "APPLICANT_PASSWORD")
+
+   // @NotNull(message = "person username is missing")
+  //  @Column(name = "APPLICANT_PASSWORD")
     private String password;
 
-    // @Column(name= "APPLICANT_FNAME")
+    // The regex below should permit only characters, but asterisk is
+    // unfortunately also valid.
+   // @Pattern(regexp = "^[\\p{L}\\p{M}*]*$", message = "Invalid characters in first name")
+    //@Size(min = 2, max = 30, message = "person first name")
+    //@Column(name= "APPLICANT_FNAME")
     private String firstName;
 
-    // @Column(name = "APPLICANT_LNAME")
+  //  @NotNull(message = "person last name is missing")
+   // @Pattern(regexp = "^[\\p{L}\\p{M}*]*$", message = "Invalid characters in persons last name")
+   // @Size(min = 2, max = 30, message = "person last name length")
+    //@Column(name = "APPLICANT_LNAME")
     private String lastName;
 
+//    @NotNull(message = "person email is missing")
+   // @Column(name = "EMAIL")
     // @Column(name = "APPLICANT_EMAIL")
     private String emailAddress;
 
+  //  @NotNull(message = "person date of birth is missing")
+   // @Column(name = "DATE_OF_BIRTH")
     // @Column(name = "APPLICANT_DATEOFBIRTH")
-    private long dateOfBirth;
+    private int dateOfBirth;
 
     /**
      * Exists only for the sake of JPA
@@ -37,9 +51,15 @@ public class Applicant implements ApplicantDTO {
 
 
     /**
-     * Constructor
+     * Creates a new instance with the specified username, password, first and last name, date of birth and email
+     * @param username The person's username
+     * @param password The person's password
+     * @param firstName The person's first name
+     * @param lastName The person's last name
+     * @param emailAddress The person's email
+     * @param dateOfBirth The person's date of birth
      */
-    public Applicant(String username, String password, String firstName, String lastName, String emailAddress, long dateOfBirth){
+    public Applicant(String username, String password, String firstName, String lastName, String emailAddress, int dateOfBirth){
         this.username = username;
         this.password = password;
         this.firstName = firstName;
@@ -70,7 +90,7 @@ public class Applicant implements ApplicantDTO {
     }
 
     @Override
-    public long getDateOfBirth() {
+    public int getDateOfBirth() {
         return dateOfBirth;
     }
